requirements.txt
python-dotenv
langchain
langcahin-community
langchain-openai
openai
pydantic